Rewrote the push command based on test experience
test/issue29.t showed places where the push flow was poor. Now we do:

    git subrepo push <subdir> [<ref>]

If there is no ref, we fetch and merge a new branch and push that. The
branch is called `subrepo-push/<subdir>` so that it won't collide with
other branches.

If there is a ref, we just push it.

@AndrewSmart, updated `test/issue29.t` to pass with new stuff. Next
commit I should be able to clean up that test code quite a bit, and use
the new push flow.