Merge #52 but change things to work like pull
Actually I refactored pull handling too.

The subrepo:pu{ll,sh} functions do the the work. They can throw errors
but shouldn't give valid output. The command:pu{ll,sh} functions do
that.

If a subrepo: can't do the normal path, we `OK=false; CODE=$num; return`
and let command deal with it. A negative CODE is not an error but a
non-standard albeit valid condition. We say something and exit 0. I
positive CODE is an error that needs special output, and we exit with
that number.

Thsi situation is not an error (imho) so we return with CODE=-1 and then
say the nice "Nothing to push here" message.