Redesign, trash the tree hash approach and use merges instead.
Tree hashes causes a lot of problems. It is very common to have
duplicate tree hashes within the repo and that will confuse the algorithm.
Instead we now try to go back to the previous approach using the
.gitrepo file containing references to the correct ancestry.
That combined with using merges will remove a lot of the problems.

Documented here: https://github.com/ingydotnet/git-subrepo/issues/142