Loosen up regex to determine if pushing to a new repo after init
The push command expects the "fatal: Couldn't find remote ref" string to be present. However, all manner of various outputs from git could fail to match that string at the beginning of the output. This relaxes this restriction with the addde bonus that the commit message is longer than the change.