Make version checking portable. #307
The fixes for #307 had a few issues. First it failed tests on OSX since
sort's -V is not portable. Found the new sort solution here:

https://stackoverflow.com/questions/4493205/unix-sort-of-version-numbers

Use cut instead of awk. Awk is pretty standard but introducing it for
this when cut is fine is too heavy.

Here are the list of commands we use besides bash and git:
```
cat
cut
dirname
grep
head
ls
mkdir
readlink
rm
tr
sort
```

We also use `date` in tests.

I'm quite comfortable with those. Anything new would require a
discussion. Note: on my box awk is 110kb and all the others are under
30.

Changed use of `[` to `[[` and other small things.