Fix regression in pull when there are no new commits
This makes the previously committed failing test case,
test/pull-no-changes.t, pass.

The regression was introduced by c03bbe5, which was itself fixing an
edge case where gitrepo wasn't updated with a new branch when the commit
was identical.  This change reverts that patch and implements a better
check condition.  Notably, the new check condition also covers the
subrepo remote needing updating, not just the subrepo branch name.

Additional tests are added to ensure I haven't broken desired behaviour
when there are no changes and there is also an alternate branch name
(with and without --update).