Add a fish shell startup script
This adds the subrepo command, but adding to the MANPATH variable in
fish seems to cause manpages to not work for other things, from
anecdotal evidence. Bash completion clearly doesn't work, but is
something that could be done for fish shell at a later point, as fish
shell completion definitions.

Since most of subrepo is executed explicitly with bash, it seems to work
fine in fish shell as far as I can tell, besides the need for a fish rc
file, as added in this commit.