Only remove .gitrepo on appropriate patches
Limit the removal of .gitrepo files to commits on the repo hosting
the subrepo. When using filter-branch and it encountered a merge
commit, continuing the filter on the second branch rewrites all
of those commits, despite that not being necessary for pre-commit.

If the rewritten commits were GPG signed the change in the hash
results in the inability to push to the remote repo.

If during development you were to change to a different branch
on the remote repo that contained a merge, you would not be able
to push upstream because the rewrite would not match the upstream
SHAs.

fixes issue #613
fixes isseu #602