Fix error tests on git 2.30 and higher
git 2.30 added a check to ensure the permissions in the .git
folder were "reasonable". When running the test in docker the
user is different than the host machine and it was throwing a
different error than expected when doing a git subrepo status
inside the .git directory. Add a configuration option to ignore
the permission mismatch during testing.