Fix subrepo init failing if defaultbranch not set
subrepo uses -e to ensure that all commands succeed, and fails
immediately if they do not. git config reports a failure if a
setting is not found causing subrepo init to silently fail if
the setting did not exist on the system. Updated to use --default
which returns a default value if the setting does not exist.

Fixes #631